8694eqn00 Cleaned RelationdataId class from redundant code, corrected creation of customer attribute
