+++ Setup/Patch/Data/CustomerRelationDataId.php
@@ -1,0 +1,89 @@
+<?php
+
+declare(strict_types=1);
+
+namespace StoreKeeper\StoreKeeper\Setup\Patch\Data;
+
+use Magento\Customer\Api\CustomerMetadataInterface;
+use Magento\Eav\Model\Config;
+use Magento\Eav\Setup\EavSetup;
+use Magento\Eav\Setup\EavSetupFactory;
+use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
+
+class CustomerRelationDataId implements DataPatchInterface
+{
+    private ModuleDataSetupInterface $moduleDataSetup;
+    private EavSetupFactory $eavSetupFactory;
+    private Config $eavConfig;
+
+    /**
+     * Construcor
+     *
+     * @param ModuleDataSetupInterface $moduleDataSetup
+     * @param EavSetupFactory $eavSetupFactory
+     * @param Config $eavConfig
+     */
+    public function __construct(
+        ModuleDataSetupInterface $moduleDataSetup,
+        EavSetupFactory $eavSetupFactory,
+        Config $eavConfig
+    ) {
+        $this->moduleDataSetup = $moduleDataSetup;
+        $this->eavSetupFactory = $eavSetupFactory;
+        $this->eavConfig = $eavConfig;
+    }
+
+    /**
+     * @return void
+     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws \Magento\Framework\Validator\ValidateException
+     */
+    public function apply()
+    {
+        $eavSetup = $this->eavSetupFactory->create(['setup' => $this->moduleDataSetup]);
+        $eavSetup->addAttribute(
+            CustomerMetadataInterface::ENTITY_TYPE_CUSTOMER,
+            'relation_data_id',
+            [
+                'label' => __('Relation Data ID'),
+                'required' => 0,
+                'position' => 200,
+                'system' => 0,
+                'user_defined' => 1,
+                'is_used_in_grid' => 1,
+                'is_visible_in_grid' => 1,
+                'is_filterable_in_grid' => 1,
+                'is_searchable_in_grid' => 1,
+            ]
+        );
+
+        $attribute = $this->eavConfig->getAttribute(
+            CustomerMetadataInterface::ENTITY_TYPE_CUSTOMER, 
+            'relation_data_id'
+        );
+
+        $attribute->setData(
+            'used_in_forms',
+            ['adminhtml_customer']
+
+        );
+        $attribute->save();
+    }
+
+    /**
+     * @return array|string[]
+     */
+    public static function getDependencies()
+    {
+        return [];
+    }
+
+    /**
+     * @return array|string[]
+     */
+    public function getAliases()
+    {
+        return [];
+    }
+}
